Close the reviewed implementation gaps and declare the remaining ones
P1/P2 of the review round: every confirmed README-vs-code mismatch is
either fixed in code or precisely declared, and the boundary topics the
review surfaced get explicit dispositions.

Code:
- load() is now true snapshot replacement per type (objects and links
  reset, no merge) and the loaded snapshot is validated against the
  model (endpoint existence, one-to-many cardinality)
- delete is RESTRICT: linked objects refuse to die (unlink first)
- primary keys cannot be modified
- traversal from a hidden origin returns nothing (no existence leak)
- MCP: aggregate_<object> tools; audit tool declared unscoped;
  buildMcpServer accepts an agent name (OO_AGENT) for stdio identity
- demo ERP adapter re-verifies its own invariant with a guarded UPDATE
- money as integer minor units; engines >= 24

README (en/ja):
- Failure semantics gains "Preconditions and freshness" and
  "Concurrent edits" declarations
- FAQ gains Foundry Functions/derived properties and idempotency/retry
  entries; audit log declared as unscoped surface; fail-closed
  "one-knob" overclaim corrected
- Non-goals: bounded context (one ontology = one context) and
  unbounded result sets
- Wording fixes: serializable -> versionable, five concepts enumerated,
  class-reflection claim softened, MCP error payload shape, article
  link, Palantir webhook citation, ja working memo removed

Tests 21 -> 27.

+++ examples/orders/ontology.ts
@@ -37,7 +37,7 @@ export const orders = defineOntology({
         id: z.string(),
         customerId: z.string(),
         status: z.enum(['pending', 'assigned', 'shipped', 'cancelled']),
-        total: z.number(),
+        total: z.number().int(), // minor units — money is not a float
         assignee: z.string().nullable(),
         // Which system of record this order lives in — write-back routes on this.
         sourceSystem: z.enum(['north', 'south']),
